+++ database/migrations/000043_library_thumbnail_dimensions.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE library DROP COLUMN thumbnail_width;
+ALTER TABLE library DROP COLUMN thumbnail_height;