+++ database/migrations/000043_library_thumbnail_dimensions.up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE library ADD COLUMN thumbnail_width INTEGER;
+ALTER TABLE library ADD COLUMN thumbnail_height INTEGER;